requirements.txt
aiohttp
chardet
aiodns
pytest
requests
asyncpg
python-dotenv
sqlalchemy
psycopg2-binary